refactor(): moved out main script file search to scriptLoading module
