+++ packages/jscad-desktop/src/app.js
@@ -6,7 +6,7 @@ const Store = require('electron-store')
 const store = new Store()
 
 const {toArray} = require('./utils')
-const {watchScript} = require('./core/scripLoading')
+const {watchScript, getScriptFile} = require('./core/scripLoading')
 const makeCsgViewer = require('../../csg-viewer/src/index')
 const loadAndDisplay = require('./ui/loadAndDisplay')
 const {setTheme} = require('./ui/setTheme')
@@ -46,51 +46,15 @@ document.getElementById('autoReload').addEventListener('click', function () {
 document.getElementById('fileLoader').addEventListener('click', function () {
   dialog.showOpenDialog({properties: ['openFile', 'openDirectory', 'multiSelections']}, function (paths) {
     console.log('loading', paths)
-    if (!paths) {
-      return
-    }
-    let mainPath = paths[0]
-    let filePath
-    const stats = fs.statSync(mainPath)
-    if (stats.isFile()) {
-      filePath = mainPath
-    } else if (stats.isDirectory()) {
-      const entries = fs.readdirSync(mainPath)
-      const acceptableMainFiles = ['main', 'index', path.parse(path.basename(mainPath)).name]
-      const jsMainFiles = acceptableMainFiles.map(x => x + '.js')
-      const jscadMainFiles = acceptableMainFiles.map(x => x + '.jscad')
-
-      const candidates = entries
-        .filter(entry => {
-          return jsMainFiles.concat(jscadMainFiles).includes(entry)
-        })
-      if (candidates.length > 0) {
-        filePath = path.join(mainPath, candidates[0])
-      }
-      // if all else fails try to use package.json to file main
-      const packageFile = path.join(mainPath, 'package.json')
-      if (fs.existsSync(packageFile)) {
-        const rMain = require(packageFile).main
-        filePath = rMain ? path.join(mainPath, rMain) : undefined
-      }
-      console.log('candites', candidates)
-        /* .reduce(function (acc, entry) {
-
-        }) */
-
-      console.log('entries', entries)
-    }
-    // stats.isDirectory()
-
-    // filePath = fileNames[0]
+    const designMainFilePath = getScriptFile(paths)
     if (settings.autoReload) {
-      watchScript(filePath, loadAndDisplay.bind(null, csgViewer))
+      watchScript(designMainFilePath, loadAndDisplay.bind(null, csgViewer))
     }
     // document.getElementById('currentFile').innerText = fileNames[0]
-    loadAndDisplay(csgViewer, filePath)
+    loadAndDisplay(csgViewer, designMainFilePath)
     // persist data
     store.set('lastDesign.name', designName)
-    store.set('lastDesign.path', filePath)
+    store.set('lastDesign.path', designMainFilePath)
   })
 })
